Centralize AI chat context state
