+++ src/apps/terminal/hooks/useTerminalLogic.ts
@@ -11,16 +11,9 @@ import { useTerminalStoreShallow } from "@/stores/helpers";
 import { useTerminalStore } from "@/stores/useTerminalStore";
 import { useLaunchApp } from "@/hooks/useLaunchApp";
 import { useAiChat } from "@/apps/chats/hooks/useAiChat";
-import { useAppStore } from "@/stores/useAppStore";
 import { useTerminalSounds } from "@/hooks/useTerminalSounds";
 import { getTextAnalytics, track } from "@/utils/analytics";
 import { useChatsStore } from "@/stores/useChatsStore";
-import { useTextEditStore } from "@/stores/useTextEditStore";
-import { useIpodStore } from "@/stores/useIpodStore";
-import { htmlToMarkdown } from "@/utils/markdown";
-import { generateHtmlFromJsonSync } from "@/utils/tiptapHtml";
-import { useInternetExplorerStore } from "@/stores/useInternetExplorerStore";
-import { useVideoStore } from "@/stores/useVideoStore";
 import { useFilesStore } from "@/stores/useFilesStore";
 import { TERMINAL_ANALYTICS } from "@/utils/analytics";
 import i18n from "@/lib/i18n";
@@ -34,235 +27,12 @@ import { commands, AVAILABLE_COMMANDS } from "../commands";
 import { helpItems } from "../index";
 import { useVimLogic } from "./useVimLogic";
 import type { AIChatMessage } from "@/types/chat";
+import { getSystemState } from "@/apps/chats/utils/systemState";
 
 interface UseTerminalLogicOptions {
   isForeground?: boolean;
   initialData?: { prefillCommand?: string };
 }
-
-// Helper function to detect user's operating system
-const detectUserOS = (): string => {
-  if (typeof navigator === "undefined") return i18n.t("apps.terminal.output.unknown");
-
-  const userAgent = navigator.userAgent;
-  const platform = navigator.platform || "";
-
-  // Check for iOS (iPhone, iPad, iPod)
-  if (
-    /iPad|iPhone|iPod/.test(userAgent) ||
-    (platform === "MacIntel" && navigator.maxTouchPoints > 1)
-  ) {
-    return "iOS";
-  }
-
-  // Check for Android
-  if (/Android/.test(userAgent)) {
-    return "Android";
-  }
-
-  // Check for Windows
-  if (/Win/.test(platform)) {
-    return "Windows";
-  }
-
-  // Check for macOS (not iOS)
-  if (/Mac/.test(platform)) {
-    return "macOS";
-  }
-
-  // Check for Linux
-  if (/Linux/.test(platform)) {
-    return "Linux";
-  }
-
-  return "Unknown";
-};
-
-// Minimal system state for AI chat requests
-const getSystemState = () => {
-  const appStore = useAppStore.getState();
-  const { username } = useChatsStore.getState();
-  const ieStore = useInternetExplorerStore.getState();
-  const videoStore = useVideoStore.getState();
-  const ipodStore = useIpodStore.getState();
-  const textEditStore = useTextEditStore.getState();
-
-  const currentVideo = videoStore.getCurrentVideo();
-  const currentTrack = ipodStore.currentSongId
-    ? ipodStore.tracks.find((t) => t.id === ipodStore.currentSongId)
-    : ipodStore.tracks[0] ?? null;
-
-  // Detect user's operating system
-  const userOS = detectUserOS();
-
-  // Use new instance-based model instead of legacy apps
-  const runningInstances = Object.entries(appStore.instances).reduce<
-    {
-      instanceId: string;
-      appId: string;
-      isForeground: boolean;
-      title?: string;
-      appletPath?: string;
-      appletId?: string;
-    }[]
-  >((acc, [instanceId, instance]) => {
-    if (!instance.isOpen) {
-      return acc;
-    }
-
-    const base = {
-      instanceId,
-      appId: instance.appId,
-      isForeground: instance.isForeground || false,
-      title: instance.title,
-    };
-    // For applet-viewer instances, include the applet path
-    if (instance.appId === "applet-viewer" && instance.initialData) {
-      const appletData = instance.initialData as {
-        path?: string;
-        shareCode?: string;
-      };
-      acc.push({
-        ...base,
-        appletPath: appletData.path || undefined,
-        appletId: appletData.shareCode || undefined,
-      });
-      return acc;
-    }
-
-    acc.push(base);
-    return acc;
-  }, []);
-
-  const foregroundInstance =
-    runningInstances.find((inst) => inst.isForeground) || null;
-  const backgroundInstances = runningInstances.filter(
-    (inst) => !inst.isForeground
-  );
-
-  // --- Local browser time information (client side) ---
-  const nowClient = new Date();
-  const userTimeZone =
-    Intl.DateTimeFormat().resolvedOptions().timeZone ||
-    i18n.t("apps.terminal.output.unknown");
-  const userTimeString = nowClient.toLocaleTimeString([], {
-    hour: "numeric",
-    minute: "2-digit",
-    hour12: true,
-  });
-  const userDateString = nowClient.toLocaleDateString([], {
-    weekday: "long",
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  });
-
-  // Convert TextEdit instances to compact markdown for prompt inclusion
-  const textEditInstances = Object.values(textEditStore.instances);
-  const textEditInstancesData = textEditInstances.map((instance) => {
-    let contentMarkdown: string | null = null;
-    if (instance.contentJson) {
-      try {
-        const htmlStr = generateHtmlFromJsonSync(instance.contentJson);
-        if (htmlStr) contentMarkdown = htmlToMarkdown(htmlStr);
-      } catch (err) {
-        console.error("Failed to convert TextEdit content to markdown:", err);
-      }
-    }
-
-    // Get title from app store instance
-    const appInstance = appStore.instances[instance.instanceId];
-    const title =
-      appInstance?.title || i18n.t("apps.terminal.output.untitled");
-
-    return {
-      instanceId: instance.instanceId,
-      filePath: instance.filePath,
-      title,
-      contentMarkdown,
-      hasUnsavedChanges: instance.hasUnsavedChanges,
-    };
-  });
-
-  // Convert IE HTML content to markdown for compact prompts
-  let ieHtmlMarkdown: string | null = null;
-  if (ieStore.aiGeneratedHtml) {
-    try {
-      ieHtmlMarkdown = htmlToMarkdown(ieStore.aiGeneratedHtml);
-    } catch (err) {
-      console.error("Failed to convert IE HTML to markdown:", err);
-    }
-  }
-
-  return {
-    username,
-    userOS,
-    userLocalTime: {
-      timeString: userTimeString,
-      dateString: userDateString,
-      timeZone: userTimeZone,
-    },
-    runningApps: {
-      foreground: foregroundInstance,
-      background: backgroundInstances,
-      instanceWindowOrder: appStore.instanceOrder,
-    },
-    internetExplorer: {
-      url: ieStore.url,
-      year: ieStore.year,
-      status: ieStore.status,
-      currentPageTitle: ieStore.currentPageTitle,
-      aiGeneratedHtml: ieStore.aiGeneratedHtml,
-      aiGeneratedMarkdown: ieHtmlMarkdown,
-    },
-    video: {
-      currentVideo: currentVideo
-        ? {
-            id: currentVideo.id,
-            url: currentVideo.url,
-            title: currentVideo.title,
-            artist: currentVideo.artist,
-          }
-        : null,
-      isPlaying: videoStore.isPlaying,
-      loopAll: videoStore.loopAll,
-      loopCurrent: videoStore.loopCurrent,
-      isShuffled: videoStore.isShuffled,
-    },
-    ipod: {
-      currentTrack: currentTrack
-        ? {
-            id: currentTrack.id,
-            url: currentTrack.url,
-            title: currentTrack.title,
-            artist: currentTrack.artist,
-          }
-        : null,
-      isPlaying: ipodStore.isPlaying,
-      loopAll: ipodStore.loopAll,
-      loopCurrent: ipodStore.loopCurrent,
-      isShuffled: ipodStore.isShuffled,
-      currentLyrics: ipodStore.currentLyrics,
-      library: ipodStore.tracks.map((t) => ({
-        id: t.id,
-        title: t.title,
-        artist: t.artist,
-      })),
-    },
-    textEdit: {
-      instances: textEditInstancesData,
-    },
-  };
-};
-
-// Helper function to check if a message is urgent (starts with "!!!!")
-export const isUrgentMessage = (content: string): boolean =>
-  content.startsWith("!!!!");
-
-// Function to clean urgent message prefix
-export const cleanUrgentPrefix = (content: string): string => {
-  return isUrgentMessage(content) ? content.slice(4).trimStart() : content;
-};
 
 export const useTerminalLogic = ({
   isForeground = true,
